Finalizacja CombatSystem dodano funkcjonujące tury usunięcie pętli while w CombatActionHandler
